Use wp_safe_redirect
`wp_safe_redirect` works the same as `wp_redirect`, but ensures the redirect will be to an allowed host—in this case, since the redirect is to an HTTPS version of the site's own URLs, that will work all of the time. It's just one additional layer of protection to prevent this redirect from being exploited.